Add C dependencies to vendor (#16127)
## What does this PR do?

This PR adds a mechanism on top of `go mod vendor` in order to copy C source files required by Beats to the folder `vendor`. From now on it is not sufficient to run `go mod vendor` as it only copies go files to `vendor`. Instead `mage vendor` or `make vendor` must be used.

If a new dependency with non-Golang sources are added, files can be moved to our vendor folder by adding a new entry to the list `copyAll` in `dev-tools/mage/gomod.go`.

```golang
CopyModules{
	Name: "github.com/godror/godror",  // name of the dependency
	FilesToCopy: []string{ // list of the folders or files to copy
		"odpi",
	},
}
```

This PR also passes `-mod=vendor` to make sure the local vendor folder is used during build.

## Why is it important?

We have two dependencies which contain C source files:

- `github.com/godror/godror`
- `github.com/tsg/go-daemon`