Fix path to mage in Docker env (#16519)
## What does this PR do?

This PR fixes the path to `mage` in Docker environment for x-pack tests.

The previously used `repo.ImportPath` included the root import path and the package subdir. Since adding the major version to the root import path, it is not correct. The new path is under `repo.CanonicalImportPath` + `repo.SubDir`.

## Why is it important?

This issue prevented X-Pack tests from running.